requirements-dev.txt
pytest==7.4.0
allure-pytest==2.13.2
flake8-bugbear==22.1.11
flake8-builtins==1.5.3
flake8-comprehensions==3.8.0
flake8-deprecated==1.3
flake8-isort==4.1.1
flake8-linenumber==0.1.8
flake8-plugin-utils==1.3.2
flake8-pytest-style==1.6.0
flake8-quotes==3.3.1
flake8==4.0.1
pyflakes==2.4.0
teamcity-messages==1.30